# Improve performance of TSI Bloom Filter