###### Required for all non-trivial PRs
- [x] Rebased/mergable
- [ ] Tests pass
- [x] CHANGELOG.md updated

This PR changes the hashing regime currently used in the TSI Bloom Filter (`murmur3`) to `xxhash`, and also, using the linear combination of two hashes trick, reduces the number of hashes generated per `Contains`/`Insert` call from two to one.

Further, allocations have been eradicated since `xxhash` is allocation-free. This allows us to remove the existing pool of hashers.

The overall improvement in performance is about **5x** .

### Note
Since this changes the hashing functions used to lookup the (possible) existence of series in TSI index files, any existing TSI files are no longer going to work. Anyone who has existing TSI indexes will need to remove them (which will put the shard back on the `inmem` index), and then recreate a TSI index using the `influx_inspect inmem2tsi` tool.

### Benchmarks
```
edd@work:~|⇒  benchstat old_bloom.txt new_bloom.txt  
name                                     old time/op    new time/op    delta
Filter_Insert/m=100_k=4_n=1000-24           258µs ± 3%      53µs ± 2%   -79.47%  (p=0.000 n=10+10)
Filter_Insert/m=1000_k=4_n=1000-24          257µs ± 3%      53µs ± 1%   -79.50%  (p=0.000 n=10+10)
Filter_Insert/m=10000_k=4_n=1000-24         255µs ± 4%      52µs ± 2%   -79.54%  (p=0.000 n=10+10)
Filter_Insert/m=100000_k=4_n=1000-24        260µs ± 2%      57µs ± 1%   -78.22%  (p=0.000 n=10+10)
Filter_Insert/m=100_k=8_n=1000-24           273µs ± 3%      65µs ± 2%   -76.17%  (p=0.000 n=10+10)
Filter_Insert/m=1000_k=8_n=1000-24          275µs ± 2%      65µs ± 2%   -76.27%  (p=0.000 n=10+10)
Filter_Insert/m=10000_k=8_n=1000-24         274µs ± 3%      65µs ± 2%   -76.26%  (p=0.000 n=10+10)
Filter_Insert/m=100000_k=8_n=1000-24        274µs ± 2%      66µs ± 1%   -76.05%  (p=0.000 n=10+9)
Filter_Insert/m=100_k=20_n=1000-24          314µs ± 2%     113µs ± 1%   -64.09%  (p=0.000 n=10+9)
Filter_Insert/m=1000_k=20_n=1000-24         315µs ± 2%     113µs ± 1%   -64.08%  (p=0.000 n=10+9)
Filter_Insert/m=10000_k=20_n=1000-24        311µs ± 0%     112µs ± 2%   -64.16%  (p=0.000 n=6+10)
Filter_Insert/m=100000_k=20_n=1000-24       317µs ± 2%     111µs ± 2%   -65.11%  (p=0.000 n=9+10)
Filter_Contains/m=100_k=4_n=1000-24         503µs ± 3%      76µs ± 2%   -84.84%  (p=0.000 n=10+10)
Filter_Contains/m=1000_k=4_n=1000-24        507µs ± 2%      78µs ± 1%   -84.58%  (p=0.000 n=9+10)
Filter_Contains/m=10000_k=4_n=1000-24       529µs ± 4%      80µs ± 1%   -84.90%  (p=0.000 n=10+10)
Filter_Contains/m=100000_k=4_n=1000-24      510µs ± 4%      76µs ± 2%   -85.17%  (p=0.000 n=9+10)
Filter_Contains/m=100_k=8_n=1000-24         526µs ± 4%     101µs ± 2%   -80.81%  (p=0.000 n=10+10)
Filter_Contains/m=1000_k=8_n=1000-24        524µs ± 4%     101µs ± 1%   -80.63%  (p=0.000 n=10+10)
Filter_Contains/m=10000_k=8_n=1000-24       531µs ± 5%      97µs ± 1%   -81.78%  (p=0.000 n=10+10)
Filter_Contains/m=100000_k=8_n=1000-24      536µs ± 2%      86µs ± 1%   -84.04%  (p=0.000 n=10+10)
Filter_Contains/m=100_k=20_n=1000-24        604µs ± 4%     137µs ± 2%   -77.38%  (p=0.000 n=10+10)
Filter_Contains/m=1000_k=20_n=1000-24       596µs ± 3%     136µs ± 1%   -77.23%  (p=0.000 n=10+10)
Filter_Contains/m=10000_k=20_n=1000-24      580µs ± 4%     125µs ± 0%   -78.51%  (p=0.000 n=10+8)
Filter_Contains/m=100000_k=20_n=1000-24     560µs ± 5%     107µs ± 1%   -80.82%  (p=0.000 n=10+8)
Filter_Merge/m=100_k=4_n=1000-24            411ns ± 3%     406ns ± 1%      ~     (p=0.265 n=10+9)
Filter_Merge/m=1000_k=4_n=1000-24           648ns ± 1%     604ns ± 1%    -6.79%  (p=0.000 n=10+10)
Filter_Merge/m=10000_k=4_n=1000-24         4.45µs ± 8%    3.95µs ± 6%   -11.31%  (p=0.000 n=10+10)
Filter_Merge/m=100000_k=4_n=1000-24        23.8µs ±21%    21.9µs ±17%      ~     (p=0.123 n=10+10)
Filter_Merge/m=100_k=8_n=1000-24            422ns ± 1%     411ns ± 1%    -2.50%  (p=0.000 n=9+9)
Filter_Merge/m=1000_k=8_n=1000-24           625ns ± 2%     614ns ± 1%    -1.78%  (p=0.001 n=10+10)
Filter_Merge/m=10000_k=8_n=1000-24         4.48µs ± 7%    3.78µs ± 5%   -15.57%  (p=0.000 n=10+10)
Filter_Merge/m=100000_k=8_n=1000-24        23.1µs ±20%    27.0µs ±11%   +16.76%  (p=0.002 n=10+10)
Filter_Merge/m=100_k=20_n=1000-24           421ns ± 1%     412ns ± 1%    -2.02%  (p=0.000 n=10+10)
Filter_Merge/m=1000_k=20_n=1000-24          625ns ± 2%     612ns ± 2%    -2.10%  (p=0.000 n=10+10)
Filter_Merge/m=10000_k=20_n=1000-24        3.51µs ± 3%    3.62µs ± 3%    +2.92%  (p=0.012 n=8+10)
Filter_Merge/m=100000_k=20_n=1000-24       26.4µs ±10%    27.0µs ±19%      ~     (p=0.549 n=9+10)

name                                     old alloc/op   new alloc/op   delta
Filter_Insert/m=100_k=4_n=1000-24          1.00kB ± 0%    0.00kB       -100.00%  (p=0.000 n=9+10)
Filter_Insert/m=1000_k=4_n=1000-24         1.00kB ± 0%    0.00kB       -100.00%  (p=0.000 n=10+10)
Filter_Insert/m=10000_k=4_n=1000-24        1.00kB ± 0%    0.00kB       -100.00%  (p=0.000 n=10+10)
Filter_Insert/m=100000_k=4_n=1000-24       1.00kB ± 0%    0.00kB       -100.00%  (p=0.000 n=10+10)
Filter_Insert/m=100_k=8_n=1000-24          1.00kB ± 0%    0.00kB       -100.00%  (p=0.000 n=10+10)
Filter_Insert/m=1000_k=8_n=1000-24         1.00kB ± 0%    0.00kB       -100.00%  (p=0.000 n=10+10)
Filter_Insert/m=10000_k=8_n=1000-24        1.00kB ± 0%    0.00kB       -100.00%  (p=0.000 n=10+10)
Filter_Insert/m=100000_k=8_n=1000-24       1.00kB ± 0%    0.00kB       -100.00%  (p=0.000 n=10+10)
Filter_Insert/m=100_k=20_n=1000-24         1.00kB ± 0%    0.00kB       -100.00%  (p=0.000 n=10+10)
Filter_Insert/m=1000_k=20_n=1000-24        1.00kB ± 0%    0.00kB       -100.00%  (p=0.000 n=10+10)
Filter_Insert/m=10000_k=20_n=1000-24       1.00kB ± 0%    0.00kB       -100.00%  (p=0.000 n=10+10)
Filter_Insert/m=100000_k=20_n=1000-24      1.00kB ± 0%    0.00kB       -100.00%  (p=0.000 n=10+10)
Filter_Contains/m=100_k=4_n=1000-24        2.00kB ± 0%    0.00kB       -100.00%  (p=0.000 n=10+10)
Filter_Contains/m=1000_k=4_n=1000-24       2.00kB ± 0%    0.00kB       -100.00%  (p=0.000 n=10+10)
Filter_Contains/m=10000_k=4_n=1000-24      2.00kB ± 0%    0.00kB       -100.00%  (p=0.000 n=10+10)
Filter_Contains/m=100000_k=4_n=1000-24     2.00kB ± 0%    0.00kB       -100.00%  (p=0.000 n=10+10)
Filter_Contains/m=100_k=8_n=1000-24        2.00kB ± 0%    0.00kB       -100.00%  (p=0.000 n=10+10)
Filter_Contains/m=1000_k=8_n=1000-24       2.00kB ± 0%    0.00kB       -100.00%  (p=0.000 n=10+10)
Filter_Contains/m=10000_k=8_n=1000-24      2.00kB ± 0%    0.00kB       -100.00%  (p=0.000 n=10+10)
Filter_Contains/m=100000_k=8_n=1000-24     2.00kB ± 0%    0.00kB       -100.00%  (p=0.000 n=10+10)
Filter_Contains/m=100_k=20_n=1000-24       2.00kB ± 0%    0.00kB       -100.00%  (p=0.000 n=10+10)
Filter_Contains/m=1000_k=20_n=1000-24      2.00kB ± 0%    0.00kB       -100.00%  (p=0.000 n=10+10)
Filter_Contains/m=10000_k=20_n=1000-24     2.00kB ± 0%    0.00kB       -100.00%  (p=0.000 n=10+10)
Filter_Contains/m=100000_k=20_n=1000-24    2.00kB ± 0%    0.00kB       -100.00%  (p=0.000 n=10+10)
Filter_Merge/m=100_k=4_n=1000-24             416B ± 0%      416B ± 0%      ~     (all equal)
Filter_Merge/m=1000_k=4_n=1000-24            416B ± 0%      416B ± 0%      ~     (all equal)
Filter_Merge/m=10000_k=4_n=1000-24           416B ± 0%      416B ± 0%      ~     (all equal)
Filter_Merge/m=100000_k=4_n=1000-24          416B ± 0%      416B ± 0%      ~     (all equal)
Filter_Merge/m=100_k=8_n=1000-24             416B ± 0%      416B ± 0%      ~     (all equal)
Filter_Merge/m=1000_k=8_n=1000-24            416B ± 0%      416B ± 0%      ~     (all equal)
Filter_Merge/m=10000_k=8_n=1000-24           416B ± 0%      416B ± 0%      ~     (all equal)
Filter_Merge/m=100000_k=8_n=1000-24          416B ± 0%      416B ± 0%      ~     (all equal)
Filter_Merge/m=100_k=20_n=1000-24            416B ± 0%      416B ± 0%      ~     (all equal)
Filter_Merge/m=1000_k=20_n=1000-24           416B ± 0%      416B ± 0%      ~     (all equal)
Filter_Merge/m=10000_k=20_n=1000-24          416B ± 0%      416B ± 0%      ~     (all equal)
Filter_Merge/m=100000_k=20_n=1000-24         416B ± 0%      416B ± 0%      ~     (all equal)
```