@e-dard 2017-09-20
/cc @pauldix regarding destructive nature of this PR on existing TSI indexes.