@pauldix 2017-09-20
@e-dard if they happen to still have old TSI files, will it show an error in the logs and ignore them?