@jwilder 2017-09-20
I noticed `Contains` is using a div and mod.  If you apply this:

```
diff --git a/pkg/bloom/bloom.go b/pkg/bloom/bloom.go
index 44031cb..22bad7a 100644
--- a/pkg/bloom/bloom.go
+++ b/pkg/bloom/bloom.go
@@ -70,7 +70,7 @@ func (f *Filter) Contains(v []byte) bool {
        h := f.hash(v)
        for i := uint64(0); i < f.k; i++ {
                loc := f.location(h, i)
-               if f.b[loc/8]&(1<<(loc%8)) == 0 {
+               if f.b[loc>>3]&(1<<(loc&7)) == 0 {
                        return false
                }
        }
```

You get a little bump in perf:

```
benchmark                                           old ns/op     new ns/op     delta
BenchmarkFilter_Insert/m=100_k=4_n=1000-8           46424         45067         -2.92%
BenchmarkFilter_Insert/m=1000_k=4_n=1000-8          46080         44299         -3.87%
BenchmarkFilter_Insert/m=10000_k=4_n=1000-8         47096         44519         -5.47%
BenchmarkFilter_Insert/m=100000_k=4_n=1000-8        45176         44210         -2.14%
BenchmarkFilter_Insert/m=100_k=8_n=1000-8           55427         53717         -3.09%
BenchmarkFilter_Insert/m=1000_k=8_n=1000-8          54040         55203         +2.15%
BenchmarkFilter_Insert/m=10000_k=8_n=1000-8         55154         55613         +0.83%
BenchmarkFilter_Insert/m=100000_k=8_n=1000-8        53395         55187         +3.36%
BenchmarkFilter_Insert/m=100_k=20_n=1000-8          83666         83027         -0.76%
BenchmarkFilter_Insert/m=1000_k=20_n=1000-8         81250         80591         -0.81%
BenchmarkFilter_Insert/m=10000_k=20_n=1000-8        83647         79664         -4.76%
BenchmarkFilter_Insert/m=100000_k=20_n=1000-8       82659         80717         -2.35%
BenchmarkFilter_Contains/m=100_k=4_n=1000-8         93512         89954         -3.80%
BenchmarkFilter_Contains/m=1000_k=4_n=1000-8        89847         89260         -0.65%
BenchmarkFilter_Contains/m=10000_k=4_n=1000-8       88685         82138         -7.38%
BenchmarkFilter_Contains/m=100000_k=4_n=1000-8      80133         81099         +1.21%
BenchmarkFilter_Contains/m=100_k=8_n=1000-8         108504        108895        +0.36%
BenchmarkFilter_Contains/m=1000_k=8_n=1000-8        111096        104599        -5.85%
BenchmarkFilter_Contains/m=10000_k=8_n=1000-8       98540         97190         -1.37%
BenchmarkFilter_Contains/m=100000_k=8_n=1000-8      91293         87978         -3.63%
BenchmarkFilter_Contains/m=100_k=20_n=1000-8        163548        153404        -6.20%
BenchmarkFilter_Contains/m=1000_k=20_n=1000-8       159283        154785        -2.82%
BenchmarkFilter_Contains/m=10000_k=20_n=1000-8      138598        130379        -5.93%
BenchmarkFilter_Contains/m=100000_k=20_n=1000-8     120709        112999        -6.39%
BenchmarkFilter_Merge/m=100_k=4_n=1000-8            66.1          65.1          -1.51%
BenchmarkFilter_Merge/m=1000_k=4_n=1000-8           201           198           -1.49%
BenchmarkFilter_Merge/m=10000_k=4_n=1000-8          2388          2282          -4.44%
BenchmarkFilter_Merge/m=100000_k=4_n=1000-8         19320         18153         -6.04%
BenchmarkFilter_Merge/m=100_k=8_n=1000-8            65.8          64.9          -1.37%
BenchmarkFilter_Merge/m=1000_k=8_n=1000-8           203           198           -2.46%
BenchmarkFilter_Merge/m=10000_k=8_n=1000-8          2466          2311          -6.29%
BenchmarkFilter_Merge/m=100000_k=8_n=1000-8         19216         18118         -5.71%
BenchmarkFilter_Merge/m=100_k=20_n=1000-8           66.8          65.8          -1.50%
BenchmarkFilter_Merge/m=1000_k=20_n=1000-8          203           198           -2.46%
BenchmarkFilter_Merge/m=10000_k=20_n=1000-8         2570          2350          -8.56%
BenchmarkFilter_Merge/m=100000_k=20_n=1000-8        19193         18533         -3.44%
```