@e-dard 2017-09-21
@jwilder see e4013f6

@jwilder @benbjohnson see be45245 for a `tsi1` versioning scheme. The idea is we'll add a `version` field the the index's `MANIFEST` (since all TSI indexes must have a `MANIFEST`). The current version will be set to `1`. When a new TSI index is created it will write out a `MANIFEST` with `"version": 1`. For existing indexes they will be deemed incompatible (since their versions will be `0` due to the absence of a `version` field on the `MANIFEST`.

Attempting to open an incompatible index results in `Open` returning an error, which is propagated up to the shard level, which itself will refuse to open. There will be plenty of log noise.

@rbetts FYI - I think it's important we make a reminder somewhere to add something to the release notes for whatever version ships TSI. We'll need to describe the index conversion steps for any users who have been running TSI on the nightlies prior to this PR change.