@bong0 2017-10-13
Thank you very much for adding this improvement, I experienced a multi-gigabyte drop in RSS memory due to the change from murmur to xxhash!

As you stated, I dropped the old indexes and recreated them and installed 1.3.7rc0 which includes this PR here :)